+++ src/utils/array.test.ts
@@ -1,7 +1,7 @@
 import {suite} from 'uvu';
 import * as t from 'uvu/assert';
 
-import {last, ensureArray} from './array.js';
+import {last, toArray} from './array.js';
 
 /* test_last */
 const test_last = suite('last');
@@ -15,15 +15,15 @@ test_last('basic behavior', () => {
 test_last.run();
 /* /test_last */
 
-/* test_ensureArray */
-const test_ensureArray = suite('ensureArray');
+/* test_toArray */
+const test_toArray = suite('toArray');
 
-test_ensureArray('basic behavior', () => {
+test_toArray('basic behavior', () => {
 	const array = [1, 2, 3];
-	t.is(array, ensureArray(array));
-	t.equal([1], ensureArray(1));
-	t.equal([{a: 1}], ensureArray({a: 1}));
+	t.is(array, toArray(array));
+	t.equal([1], toArray(1));
+	t.equal([{a: 1}], toArray({a: 1}));
 });
 
-test_ensureArray.run();
-/* /test_ensureArray */
+test_toArray.run();
+/* /test_toArray */