+++ src/project/dist.task.ts
@@ -3,7 +3,6 @@ import {copy} from '../fs/node.js';
 import {paths, toBuildOutPath} from '../paths.js';
 import {isTestBuildFile, isTestBuildArtifact} from '../fs/testModule.js';
 import {printPath} from '../utils/print.js';
-import {clean} from '../fs/clean.js';
 import {loadGroConfig} from '../config/config.js';
 import {configureLogLevel} from '../utils/log.js';
 import {printBuildConfig} from '../config/buildConfig.js';
@@ -12,8 +11,6 @@ export const task: Task = {
 	description: 'create and link the distribution',
 	dev: false,
 	run: async ({invokeTask, dev, log}) => {
-		await clean({dist: true}, log);
-
 		// This reads the `dist` flag on the build configs to help construct the final dist directory.
 		// See the docs at `./docs/config.md`.
 		const config = await loadGroConfig(dev);
@@ -27,7 +24,10 @@ export const task: Task = {
 						? paths.dist
 						: `${paths.dist}${printBuildConfig(buildConfig)}`;
 				log.info(`copying ${printPath(buildOutDir)} to ${printPath(distOutDir)}`);
-				return copy(buildOutDir, distOutDir, {filter: (id) => isDistFile(id)});
+				return copy(buildOutDir, distOutDir, {
+					overwrite: false, // let the TypeScript output take priority, but allow other files like Svelte
+					filter: (id) => isDistFile(id),
+				});
 			}),
 		);