+++ changelog.md
@@ -5,8 +5,9 @@
 - **break**: rename `src/fs/node.ts` from `src/fs/nodeFs.ts`
   ([#154](https://github.com/feltcoop/gro/pull/154))
 - **break**: rename `toArray` from `ensureArray`
-  ([#154](https://github.com/feltcoop/gro/pull/154))
-- export `AsyncStatus` from root
+  ([#117](https://github.com/feltcoop/gro/pull/117))
+- export many more things from root: `import {/* !!! */} from '@feltcoop/gro';`
+  ([#117](https://github.com/feltcoop/gro/pull/117))
 
 ## 0.13.1